FastAPI/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
fastapi
azure-storage-blob
python-multipart
PyPDF2
llama_index
openai
python-dotenv
uvicorn
sqlalchemy
